Bugfixes for process pool executor calculations
`calculate_arbitrage_return_best` was throwing `AttributeError` exceptions about a missing `_state_lock` attribute during serialization. The pool was being re-serialized after the state lock was stripped, causing the AttributeError. The `swap_pools` key in `self.best` was causing those objects to be pulled into the serializer, but they are unnecessary and have been removed.

`self.stable_swap` was not included in `__slots__`, so was similarly not being found when a `CamelotLiquidityPool` was used inside a process pool.